requirements.txt
## Tests & Releng
setuptools-scm
pytest
pytest-xdist
pytest-cov
coverage 

## Requirements
argcomplete
colorama
python-gnupg
requests
